last commit

+++ new/shorten/views.py
@@ -1,12 +1,18 @@
 from django.shortcuts import render, get_object_or_404
 from django.http import HttpResponse, HttpResponseRedirect
 
+#To use it as dynamic urls
+from django.urls import reverse
+
+from django.contrib.auth.decorators import login_required
+
+
 from .models import bitly
 from .forms import bitlyform, editBitly
 # Create your views here.
 
 def index(request):
-    objects = bitly.objects.all()
+    objects = bitly.objects.all()[::-1]
     print(objects)
 
     context = {'objs' : objects}
@@ -22,7 +28,7 @@ def create(request):
         instance.datewise = "{}"
         instance.save()
 
-        return HttpResponseRedirect("http://127.0.0.1:8000/home/")
+        return HttpResponseRedirect(reverse('index'))
 
     context = {"urlform": form}
     return render(request, "create.html", context)
@@ -42,19 +48,23 @@ def goto(request, shortcode=None):
         qs.save()
     return HttpResponseRedirect(qs.long_url)
 
+
 def update(request, pk=None):
-    qs = get_object_or_404(bitly, id=pk)
-    form = editBitly(request.POST or None, instance=qs)
+    if request.user.is_authenticated:
+        qs = get_object_or_404(bitly, id=pk)
+        form = editBitly(request.POST or None, instance=qs)
 
-    if form.is_valid():
-        form.save()
-        return HttpResponseRedirect("http://127.0.0.1:8000/home")
+        if form.is_valid():
+            form.save()
+            return HttpResponseRedirect(reverse('index'))
 
-    context = {'urlform': form}
-    return render(request, 'create.html', context)
+        context = {'urlform': form}
+        return render(request, 'create.html', context)
+    return HttpResponseRedirect(reverse('index'))
 
 def delete(request, pk=None):
-    qs = get_object_or_404(bitly, id=pk)
-    qs.delete()
-    return HttpResponseRedirect("http://127.0.0.1:8000/home")
+    if request.user.is_authenticated:
+        qs = get_object_or_404(bitly, id=pk)
+        qs.delete()
+    return HttpResponseRedirect(reverse('index'))
